% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvirOmix-package.R
\docType{package}
\name{EnvirOmix-package}
\alias{EnvirOmix}
\alias{EnvirOmix-package}
\title{EnvirOmix: Integrating Environmental and Multiomic Data for Precision Health}
\description{
This is the companion package for the book "Integrating Multi-Omics with Environmental data for Precision Health". It contains all of the functions and data required to run all of the examples in the book.
}
\author{
\strong{Maintainer}: Jesse Goodrich \email{jagoodri@usc.edu} (\href{https://orcid.org/0000-0001-6615-0472}{ORCID})

Authors:
\itemize{
  \item Hongxu Wang
  \item Qiran Jia
  \item David Conti
  \item Shardul Nazirkar
}

}
\keyword{internal}
