% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidimum.R
\name{hidimum}
\alias{hidimum}
\title{High Dimensional Multiomic Mediation (hidimum)}
\usage{
hidimum(
  exposure,
  outcome,
  omics_lst,
  covs,
  Y.family = "binomial",
  M.family = "gaussian",
  integration,
  n_boot
)
}
\arguments{
\item{exposure}{A numeric vector for the exposure variable}

\item{outcome}{A numeric vector for the outcome variable}

\item{covs}{A numeric matrix representing the covariates}

\item{n_boot}{number indicating number of bootstrap estimates to perform
for calculating the se of the mediation effect (only applies for
intermediate integration, otherwise, ignored)}

\item{omics}{A list of numeric matrices representing omics data}
}
\value{
A tidy dataframe summarizing the results of HIMA analysis
}
\description{
Given exposure, outcome and multiple omics data, this
function runs high dimensional mediation with early, intermediate, or
late integration of the multiomics datasets.#'
}
