% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafamum.R
\name{lafamum}
\alias{lafamum}
\title{Latent Factor Multiomic Mediation (lafamum)}
\usage{
lafamum(
  exposure,
  outcome,
  omics_lst,
  covs,
  Y.family = "gaussian",
  fdr.level = 0.05,
  integration,
  jive.rankJ = NULL,
  jive.rankA = NULL
)
}
\arguments{
\item{exposure}{A numeric vector for the exposure variable}

\item{outcome}{A numeric vector for the outcome variable}

\item{omics_lst}{A list of numeric matrices representing omics data}

\item{covs}{A numeric matrix representing the covariates}

\item{Y.family}{A character string indicating the family of the outcome}

\item{fdr.level}{The FDR level for selecting significant latent factors}

\item{integration}{A character string indicating the integration method (one
of early, intermediate, or late). For early and late integration,
dimensionality reduction is performed using PCA analysis; for intermediate
integration, dimensionality reduction is performed using Joint and
Individual Variance Explained (JIVE).}

\item{jive.rankJ}{If integration = "intermediate", number of joint factors
for JIVE to estimate. If NULL, then jive estimates the optimum number of
joint factors.}

\item{jive.rankA}{If integration = "intermediate", number of individual
factors for JIVE to estimate. If NULL, then jive estimates this variable.
If specified, should be a numeric vector with the same length as
dim(omics_lst).}
}
\value{
A list including two dataframes summarizing the results of
HIMA analysis. One vector indicating integration type, one dataframe
includes the significant result of HIMA analysis with latent factors as
mediators, and one dataframe includes the result of individual feature
correlations with significant latent factors.
}
\description{
Given exposure, outcome and multiple omics data, runs mediation with latent
factors defined by multiomic data. Performs a two step analysis, where first
the multiomics data goes through a dimensionality reduction step,
followed by a high dimensional mediation analysis using HIMA.
For early and late integration, dimensionality reduction is performed
using principal component analysis (PCA) by selecting the top i principal
components which explained >80\% of the variance. For intermediate
integration, dimensionality reduction is performed using Joint and
Individual Variance Explained (JIVE).
}
