% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hidimum.R
\name{plot_hidimum}
\alias{plot_hidimum}
\title{Plot for the High Dimensional Multiomic Mediation Analysis}
\usage{
plot_hidimum(result_hima)
}
\arguments{
\item{result_hima}{an output dataframe from the hidimum analysis}
}
\value{
a ggplot figure.
}
\description{
Given a tidy dataframe summarizing the results of HIMA analysis
}
