% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lafamum.R
\name{plot_lafamum}
\alias{plot_lafamum}
\title{Plot result of mediation analysis with latent factors}
\usage{
plot_lafamum(lafamum_res)
}
\arguments{
\item{lafamum_res}{A list including two tidy dataframes
summarizing the results of HIMA analysis
and one vector indicating integration type}
}
\value{
a figure of the result of mediation with latent factors in given integration
}
\description{
Given a list including two tidy dataframes and one integration type vector
summarizing  mediation analysis result
function runs plotting and returns a plot
}
