% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated HELIX dataset}
\format{
An object of class \code{list} of length 6.
}
\source{
\url{http://www.projecthelix.eu/}
}
\usage{
simulated_data
}
\description{
A simulated dataset that mimics the structure of the Human Early Life
Exposome (HELIX) dataset.
}
\details{
A list of 6:
\describe{
\item{phenotype}{
A data frame with 420 observations of 6 variables:
\describe{
\item{id}{Character vector of IDs}
\item{hs_child_age_yrs_None}{Numerical vector of child ages}
\item{hs_hg_m_scaled}{Numerical vector of scaled hs_hg_m values}
\item{ck18_scaled}{Numerical vector of scaled ck18 values}
\item{e3_sex_None}{Factor with 2 levels ("female", "male")}
\item{h_fish_preg_Ter}{Factor with 3 levels ("1", "2", "3")}
}
}
\item{methylome}{
A numerical matrix of dimensions 420x28, with row names representing individual
IDs and column names representing methylation sites.
}
\item{transcriptome}{
A numerical matrix of dimensions 420x28, with row names representing individual
IDs and column names representing transcript sites.
}
\item{miRNA}{
A numerical matrix of dimensions 420x28, with row names representing individual
IDs and column names representing miRNA sites.
}
\item{proteome}{
A numerical matrix of dimensions 420x28, with row names representing individual
IDs and column names representing proteins.
}
\item{metabolome}{
A numerical matrix of dimensions 420x28, with row names representing individual
IDs and column names representing metabolites.
}
}
}
\keyword{datasets}
