% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidimum.R
\name{hidimum}
\alias{hidimum}
\title{High Dimensional Multiomic Mediation (hidimum)}
\usage{
hidimum(
  exposure,
  outcome,
  omics_lst,
  covs,
  Y.family = "binomial",
  M.family = "gaussian",
  integration,
  n_boot,
  bh.fdr = 0.05,
  ...,
  n_cores = NULL
)
}
\arguments{
\item{exposure}{A numeric vector for the exposure variable}

\item{outcome}{A numeric vector for the outcome variable}

\item{omics_lst}{A list of numeric matrices representing omics data}

\item{covs}{A numeric matrix representing the covariates}

\item{Y.family}{A character string indicating the family of the outcome}

\item{M.family}{A character string indicating the family of the mediator}

\item{integration}{A character string indicating the integration method (one
of early, intermediate, or late)}

\item{n_boot}{number indicating number of bootstrap estimates to perform
for calculating the se of the mediation effect (only applies for
intermediate integration, otherwise, ignored)}

\item{bh.fdr}{Bonferonni-Hochberg FDR correction threshold for selecting
significant features from HIMA for early and late integration. Defaults
to 0.05.}

\item{n_cores}{Optional, number of cores for parallelization for
bootstrapping for intermediate integration. If unspecified and
integration = intermediate, defaults to parallel::detectCores().}
}
\value{
A tidy dataframe summarizing the results of HIMA analysis
}
\description{
Given exposure, outcome and multiple omics data, this
function runs high dimensional mediation with early, intermediate, or
late integration of the multiomics datasets.#'
}
