% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hidimum.R
\name{plot_hidimum}
\alias{plot_hidimum}
\title{Plot for the High Dimensional Multiomic Mediation Analysis}
\usage{
plot_hidimum(result_hidimum)
}
\arguments{
\item{result_hidimum}{an output dataframe from the hidimum analysis}
}
\value{
a figure of the result of high dimensinal multiomic mediation analysis in given integration
}
\description{
Given a tidy dataframe summarizing the results of HIMA analysis
}
\examples{
# Load Example Data
 data("simulated_data")
 covars <- c("e3_sex_None", "hs_child_age_yrs_None")

 # Extract exposure and outcome data
 exposure <- simulated_data[["phenotype"]]$hs_hg_m_scaled
 outcome  <- simulated_data[["phenotype"]]$ck18_scaled

 # Get numeric matrix of covariates
 covs <- simulated_data[["phenotype"]][covars]
 covs$e3_sex_None <- ifelse(covs$e3_sex_None == "male", 1, 0)

 # create list of omics data
 omics_lst <- simulated_data[-which(names(simulated_data) == "phenotype")]

 # High Dimensional Multiomic Mediation with Early integration
 result_hidimum_early <- hidimum(exposure = exposure,
                                 outcome = outcome,
                                 omics_lst = omics_lst,
                                 covs = covs,
                                 Y.family = "gaussian",
                                 M.family = "gaussian",
                                 integration = "early")
 # Plot result
 fig_early <- plot_hidimum(result_hidimum_early)

}
